+++ apps/web/src/features/auth/use-auth.ts
@@ -1,0 +1,14 @@
+"use client";
+
+import { AuthContext } from "@/features/auth/auth-provider";
+import { useContext } from "react";
+
+export function useAuth() {
+  const context = useContext(AuthContext);
+
+  if (!context) {
+    throw new Error("useAuth must be used within AuthProvider");
+  }
+
+  return context;
+}